Create components + tests folders + organise file
